Cleanup CSS transitions code, move shared code to another place
