+++ packages/react-native-reanimated/src/css/platform/native/normalization/animation/shorthand.ts
@@ -1,0 +1,6 @@
+'use strict';
+
+export function parseCSSAnimationShorthand(value: string) {
+  const parts = value.split(',');
+  return parts.map((part) => part.trim());
+}
