+++ packages/react-native-reanimated/src/css/platform/native/normalization/common/index.ts
@@ -4,3 +4,4 @@ export {
   normalizeDuration,
   normalizeTimingFunction,
 } from './settings';
+export * from './shorthand';
