+++ packages/react-native-reanimated/src/css/platform/native/normalization/transition/config.ts
@@ -21,8 +21,8 @@ import {
   normalizeTimingFunction,
 } from '../common';
 import { normalizeTransitionBehavior } from './settings';
-import type { ExpandedConfigProperties } from './shorthand';
-import { parseTransitionShorthand } from './shorthand';
+import type { ExpandedCSSTransitionConfigProperties } from './shorthand';
+import { parseCSSTransitionShorthand } from './shorthand';
 
 export const ERROR_MESSAGES = {
   invalidTransitionProperty: (
@@ -32,13 +32,13 @@ export const ERROR_MESSAGES = {
 
 function getExpandedConfigProperties(
   config: CSSTransitionProperties
-): ExpandedConfigProperties {
+): ExpandedCSSTransitionConfigProperties {
   const configEntries = Object.entries(config);
   const shorthandIndex = config.transition
     ? configEntries.findIndex(([key]) => key === 'transition')
     : -1;
   const result: AnyRecord = config.transition
-    ? parseTransitionShorthand(config.transition)
+    ? parseCSSTransitionShorthand(config.transition)
     : {};
   // If there is a shorthand `transition` property, all properties specified
   // before are ignored and only these specified later are taken into account
@@ -51,11 +51,11 @@ function getExpandedConfigProperties(
     result[key] = convertPropertyToArray(value);
   }
 
-  return result as ExpandedConfigProperties;
+  return result as ExpandedCSSTransitionConfigProperties;
 }
 
 const hasTransitionProperties = (
-  transitionProperty: ExpandedConfigProperties['transitionProperty']
+  transitionProperty: ExpandedCSSTransitionConfigProperties['transitionProperty']
 ): transitionProperty is string[] =>
   !!transitionProperty?.length &&
   transitionProperty.some((prop) => prop !== 'none');
